# Speed up decoding HTML Entities.