--- a/lib/common.js
+++ b/lib/common.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import encode from "mdurl/encode.js";
-import { decodeHTML } from "entities";
+import { decodeEntity } from "html-entities";
 
 var C_BACKSLASH = 92;
 
@@ -58,7 +58,7 @@ var unescapeChar = function(s) {
     if (s.charCodeAt(0) === C_BACKSLASH) {
         return s.charAt(1);
     } else {
-        return decodeHTML(s);
+        return decodeEntity(s);
     }
 };