--- a/lib/inlines.js
+++ b/lib/inlines.js
@@ -3,7 +3,7 @@
 import Node from "./node.js";
 import * as common from "./common.js";
 import fromCodePoint from "./from-code-point.js";
-import { decodeHTML } from "entities";
+import { decodeEntity } from "html-entities";
 import "string.prototype.repeat"; // Polyfill for String.prototype.repeat
 
 var normalizeURI = common.normalizeURI;
@@ -756,7 +756,7 @@ var removeBracket = function() {
 var parseEntity = function(block) {
     var m;
     if ((m = this.match(reEntityHere))) {
-        block.appendChild(text(decodeHTML(m)));
+        block.appendChild(text(decodeEntity(m)));
         return true;
     } else {
         return false;