Improve performance by using `decodeEntity()` from `html-entities` when decoding entities.

## Benchmark results

### Before

```sh
$ node bench/bench.js bench/samples/inline-entity.md
commonmark.js x 35,682 ops/sec ±1.52% (87 runs sampled)
showdown.js x 7,871 ops/sec ±4.67% (84 runs sampled)
marked.js x 55,382 ops/sec ±1.05% (96 runs sampled)
markdown-it x 41,105 ops/sec ±20.31% (89 runs sampled)
```

### After

```sh
$ node bench/bench.js bench/samples/inline-entity.md
commonmark.js x 42,870 ops/sec ±5.41% (83 runs sampled)
showdown.js x 8,168 ops/sec ±1.11% (89 runs sampled)
marked.js x 51,986 ops/sec ±4.46% (86 runs sampled)
markdown-it x 47,063 ops/sec ±3.06% (94 runs sampled)
```