@jgm 2021-03-29
Thanks! I like the speed improvement, but I'm less sure about other tradeoffs in switching from `entities` to `html-entities`.  Overall quality of the libraries, degree of usage, unpacked size?  I'm not a JavaScript person so hopefully others can also weigh in on this.