@jgm 2021-03-29
Yes, but `entities` has 27k installations per week, so it's the more widely used library, and it reports a much smaller unpacked size (57k vs 161k).  That's why the decision doesn't seem obvious to me.