@mdevils 2021-03-29
@jgm the latest `html-entities` is `99.5 kB` :)